hercules/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
supabase-py
